change float to double

+++ app/src/main/java/com/satryaway/paypaychallenge/models/LiveModel.kt
@@ -3,5 +3,5 @@ package com.satryaway.paypaychallenge.models
 data class LiveModel (
     val success: Boolean?,
     val source: String?,
-    val quotes: HashMap<String, Float>?
+    val quotes: HashMap<String, Double>?
 )
